third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/anchor-scroll-position-try-007-expected.txt

This is a testharness.js-based test.
[FAIL] Should use the second position option with enough space below (and not enough above)
  assert_equals: Anchored element should be at the right of anchor expected 135 but got -65
Harness: the test ran to completion.
